Follow file paths without extension
When a file path under the cursor does not have an extension (which is
quite common in `import` statements in JavaScript projects), Vim will
now try to append it so the `gf` command can function properly.

This has only been applied to JavaScript and TypeScript files (including
their JSX flavours).